Better handling for reusable settings in dc

+++ python/src/mdedevicecontrol/dcgraph.py
@@ -234,6 +234,9 @@ class Graph:
 
     async def delete_device_configuration(self,policy_id):
 
+        if policy_id is None:
+            logger.error("policy_id is None")
+            return RuntimeError()
         logger.debug("policy id="+policy_id)
         try:
             result = await self.graph_client.device_management.device_configurations.by_device_configuration_id(policy_id).delete()
@@ -411,7 +414,7 @@ class Graph:
         setting.setting_definition_id = "device_vendor_msft_defender_configuration_devicecontrol_policygroups_{groupid}_groupdata"
         
         logger.debug("Update Group v2 "+str(group))
-        result = await self.graph_client.device_management.reusable_policy_settings.by_device_management_reusable_policy_setting_id(group_id).put(setting)
+        result = await self.graph_client.device_management.reusable_policy_settings.by_device_management_reusable_policy_setting_id(group_id).put(setting) 
         logger.debug(str(result))
         return result
